# .mem/openapi.yaml
openapi: 3.0.3
info:
  title: Co-Work API
  description: API documentation for the Co-Work collaborative project management system
  version: 1.0.0
  contact:
    name: Co-Work Team
servers:
  - url: /api
    description: API base path
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      required:
        - error
...
    Version:
      type: object
      properties:
        id:
          type: integer
          description: Version ID
        fileId:
          type: integer
          description: File ID
        title:
          type: string
          description: Version title
        author:
          type: string
          description: Version author
        content:
          type: string
          description: Version content
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - fileId
        - title
        - author
        - content
        - createdAt
    Task:
      type: object
      properties:
        id:
          type: integer
          description: Task ID
        fileId:
          type: integer
          description: File ID
        title:
          type: string
          description: Task title
        status:
          type: string
          enum: [pending, in_progress, completed]
          description: Task status
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - fileId
        - title
        - status
        - createdAt
        - updatedAt
paths:
...
  /versions:
    get:
      tags:
        - Versions
      summary: List file versions
      description: Retrieve version history for a file
      parameters:
        - name: id
          in: query
          description: Version ID for single version retrieval
          schema:
            type: integer
        - name: fileId
          in: query
          description: File ID (required for listing versions)
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Version'
                  - type: array
                    items:
                      $ref: '#/components/schemas/Version'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Versions
      summary: Create file version
      description: Create a new version of a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fileId
                - title
                - author
                - content
              properties:
                fileId:
                  type: integer
                  description: File ID
                title:
                  type: string
                  description: Version title
                author:
                  type: string
                  description: Version author
                content:
                  type: string
                  description: Version content
      responses:
        '201':
          description: Version created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Versions
      summary: Delete file version
      description: Delete a version from history
      parameters:
        - name: id
          in: query
          required: true
          description: Version ID
          schema:
            type: integer
      responses:
        '200':
          description: Version deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  version:
                    $ref: '#/components/schemas/Version'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks:
    get:
      tags:
        - Tasks
      summary: List tasks
      description: Retrieve tasks for a file
      parameters:
        - name: id
          in: query
          description: Task ID for single task retrieval
          schema:
            type: integer
        - name: fileId
          in: query
          description: File ID (required for listing tasks)
          schema:
            type: integer
        - name: status
          in: query
          description: Filter by task status
          schema:
            type: string
            enum: [pending, in_progress, completed]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Task'
                  - type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Tasks
      summary: Create task
      description: Create a new task for a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fileId
                - title
              properties:
                fileId:
                  type: integer
                  description: File ID
                title:
                  type: string
                  description: Task title
                status:
                  type: string
                  enum: [pending, in_progress, completed]
                  default: pending
                  description: Task status
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Tasks
      summary: Update task
      description: Update an existing task
      parameters:
        - name: id
          in: query
          required: true
          description: Task ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Task title
                status:
                  type: string
                  enum: [pending, in_progress, completed]
                  description: Task status
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete an existing task
      parameters:
        - name: id
          in: query
          required: true
          description: Task ID
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  task:
                    $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'